# src/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Gateway Management API",
    "description": "Documentation automatically generated by the <b>swagger-autogen</b> module."
  },
  "host": "localhost:3000",
  "basePath": "/",
  "tags": [
    {
      "name": "Auth",
      "description": "Endpoints"
    },
    {
      "name": "Gateway",
      "description": "Endpoints"
    }
  ],
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "bearerAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization",
      "description": "Enter the token with the `Bearer: ` prefix, e.g. \"Bearer abcde12345\"."
    },
    "basicAuth": {
      "type": "basic"
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/api/gateway/{id}": {
      "get": {
        "tags": [
          "Gateway"
        ],
        "description": "Endpoint to show an specific gateway by it",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID of the gateway"
          }
        ],
        "responses": {
          "200": {
            "description": "Gateway found.",
            "schema": {
              "$ref": "#/definitions/Gateway"
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "patch": {
        "tags": [
          "Gateway"
        ],
        "description": "Endpoint to update an specific gateway by it",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "ID of the gateway"
          },
          {
            "name": "updatedGateway",
            "in": "body",
            "description": "Gateway field to be updated information.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UpdateGateway"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Gateway found.",
            "schema": {
              "$ref": "#/definitions/Gateway"
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Gateway"
        ],
        "description": "Endpoint to delete an specific gateway by it",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the gateway to be deleted",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {},
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/gateway": {
      "get": {
        "tags": [
          "Gateway"
        ],
        "description": "Endpoint to list all gateways",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "Gateway"
        ],
        "description": "Endpoint to create a gateway",
        "parameters": [
          {
            "name": "newGateway",
            "in": "body",
            "description": "Gateway information.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AddGateway"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/gateway/{id}/device": {
      "post": {
        "tags": [
          "Gateway"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "newGateway",
            "in": "body",
            "description": "Gateway information.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AddDevice"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/gateway/{id}/device/{deviceId}": {
      "delete": {
        "tags": [
          "Gateway"
        ],
        "description": "Endpoint to delete an specific gateway by it",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the gateway",
            "required": true,
            "type": "string"
          },
          {
            "name": "deviceId",
            "in": "path",
            "description": "ID of the device to be deleted",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {},
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/signup": {
      "post": {
        "tags": [
          "Auth"
        ],
        "description": "Signup a user",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Person"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/login": {
      "get": {
        "tags": [
          "Auth"
        ],
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "basicAuth": []
          }
        ]
      }
    }
  },
  "definitions": {
    "Person": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "Admin"
        },
        "userName": {
          "type": "string",
          "example": "admin"
        },
        "password": {
          "type": "string",
          "example": "P455w0rd*"
        }
      }
    },
    "Gateway": {
      "type": "object",
      "properties": {
        "serialNumber": {
          "type": "string",
          "example": "M720"
        },
        "name": {
          "type": "string",
          "example": "Example gateway"
        },
        "ipv4": {
          "type": "string",
          "example": "192.168.1.1"
        },
        "devices": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "uid": {
                "type": "string",
                "example": "2111"
              },
              "vendor": {
                "type": "string",
                "example": "somevendor"
              },
              "status": {
                "type": "string",
                "example": "offline"
              }
            }
          }
        }
      }
    },
    "AddGateway": {
      "type": "object",
      "properties": {
        "serialNumber": {
          "type": "string",
          "example": "M720"
        },
        "name": {
          "type": "string",
          "example": "Example gateway"
        },
        "ipv4": {
          "type": "string",
          "example": "192.168.1.1"
        },
        "devices": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "uid": {
                "type": "string",
                "example": "2110"
              },
              "vendor": {
                "type": "string",
                "example": "somevendor"
              },
              "status": {
                "type": "string",
                "example": "online"
              }
            }
          }
        }
      }
    },
    "UpdateGateway": {
      "type": "object",
      "properties": {
        "serialNumber": {
          "type": "string",
          "example": "number update"
        },
        "name": {
          "type": "string",
          "example": "name update"
        },
        "ipv4": {
          "type": "string",
          "example": "192.168.1.2"
        }
      }
    },
    "AddDevice": {
      "type": "object",
      "properties": {
        "uid": {
          "type": "string",
          "example": "2110"
        },
        "vendor": {
          "type": "string",
          "example": "somevendor"
        },
        "status": {
          "type": "string",
          "example": "online"
        }
      }
    }
  }
}